Extract API source resolver
Move shared API source resolution out of ApiCommand and into ApiSourceResolver, then update type, member, and source commands to consume the service. Reset member parser test preprocessor state so direct parser tests match production preprocessing behavior.
